add billing block for all routes

+++ main/views.py
@@ -76,7 +76,11 @@ def index(request):
         return render(request, "main/landing.html")
 
 
+@require_safe
+@login_required
 def people(request):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
     if request.user.is_authenticated:
         if request.user.profile.is_admin and not request.user.profile.stripe_id:
             return redirect("main:billing_setup")
@@ -188,6 +192,8 @@ def logout(request):
 @require_http_methods(["HEAD", "GET", "POST"])
 @login_required
 def invite(request):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
     company = Company.objects.get(route=request.user.profile.company.route)
     UserFormSet = formset_factory(UserForm, max_num=100, extra=100)
     if request.method == "POST":
@@ -261,6 +267,9 @@ def invite_open_refresh(request):
 @require_http_methods(["HEAD", "GET", "POST"])
 @login_required
 def invite_setup(request):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     if request.user.profile.name:
         return redirect("main:profile", request.user.username)
 
@@ -414,6 +423,9 @@ def subscribe_thanks(request):
 @require_http_methods(["HEAD", "GET", "POST"])
 @login_required
 def profile(request, username):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     if not username:
         company = Company.objects.get(route=request.user.profile.company.route)
         return render(request, "main/profile.html", {"company": company})
@@ -494,6 +506,9 @@ def company_logo(request):
 @require_http_methods(["HEAD", "GET", "POST"])
 @login_required
 def settings_user(request):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     if request.method == "POST":
         form = ProfileForm(
             request.POST,
@@ -528,6 +543,9 @@ def settings_user(request):
 @require_http_methods(["HEAD", "GET", "POST"])
 @login_required
 def settings_company(request):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     if not request.user.profile.is_admin:
         return redirect("main:settings_user")
 
@@ -546,6 +564,9 @@ def settings_company(request):
 @require_http_methods(["HEAD", "GET", "POST"])
 @login_required
 def resources(request):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     companytags = (
         CompanyTag.objects.all()
         .filter(company=request.user.profile.company)
@@ -569,6 +590,9 @@ def resources(request):
 @require_http_methods(["HEAD", "GET", "POST"])
 @login_required
 def resources_view(request, resource_slug):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     resource = get_object_or_404(Resource, slug=resource_slug)
     return render(request, "main/resources_view.html", {"resource": resource})
 
@@ -576,6 +600,9 @@ def resources_view(request, resource_slug):
 @require_http_methods(["HEAD", "GET", "POST"])
 @login_required
 def resources_create(request):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     if request.method == "POST":
         form = ResourceForm(request.POST)
         if form.is_valid():
@@ -606,6 +633,9 @@ def resources_create(request):
 @require_http_methods(["HEAD", "GET", "POST"])
 @login_required
 def resources_edit(request, resource_slug):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     resource = Resource.objects.get(slug=resource_slug)
     if request.method == "POST":
         form = ResourceForm(
@@ -643,6 +673,9 @@ def resources_edit(request, resource_slug):
 @require_http_methods(["HEAD", "POST"])
 @login_required
 def resources_delete(request, resource_slug):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     if request.method == "POST":
         resource = Resource.objects.get(slug=resource_slug)
         form = DeleteResourceForm(request.POST, instance=resource)
@@ -656,6 +689,9 @@ def resources_delete(request, resource_slug):
 
 
 def resources_pins(request):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     companytags = (
         CompanyTag.objects.all()
         .filter(company=request.user.profile.company)
@@ -682,6 +718,9 @@ def resources_pins(request):
 @require_http_methods(["HEAD", "GET", "POST"])
 @login_required
 def questions(request):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     questions = Question.objects.filter(company=request.user.profile.company).order_by(
         "-updated_at"
     )
@@ -691,6 +730,9 @@ def questions(request):
 @require_http_methods(["HEAD", "GET", "POST"])
 @login_required
 def questions_view(request, question_slug):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     question = Question.objects.get(slug=question_slug)
     if request.method == "POST":
         form = AnswerForm(request.POST)
@@ -734,6 +776,9 @@ def questions_view(request, question_slug):
 @require_http_methods(["HEAD", "GET", "POST"])
 @login_required
 def questions_create(request):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     if request.method == "POST":
         form = QuestionForm(request.POST)
         if form.is_valid():
@@ -758,6 +803,9 @@ def questions_create(request):
 @require_http_methods(["HEAD", "GET", "POST"])
 @login_required
 def questions_edit(request, question_slug):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     question = Question.objects.get(slug=question_slug)
     if request.user != question.author:
         return redirect("main:questions")
@@ -786,6 +834,9 @@ def questions_edit(request, question_slug):
 @require_http_methods(["POST"])
 @login_required
 def questions_delete(request, question_slug):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     if request.method == "POST":
         question = Question.objects.get(slug=question_slug)
         if request.user != question.author:
@@ -803,6 +854,9 @@ def questions_delete(request, question_slug):
 @require_http_methods(["HEAD", "GET", "POST"])
 @login_required
 def questions_edit_answer(request, question_slug, answer_id):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     question = Question.objects.get(slug=question_slug)
     answer = Answer.objects.get(id=answer_id)
     if request.user != answer.author or question != answer.question:
@@ -826,6 +880,9 @@ def questions_edit_answer(request, question_slug, answer_id):
 @require_http_methods(["POST"])
 @login_required
 def questions_delete_answer(request, question_slug, answer_id):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     if request.method == "POST":
         question = Question.objects.get(slug=question_slug)
         answer = Answer.objects.get(id=answer_id)
@@ -896,6 +953,9 @@ def privacy(request):
 @require_http_methods(["HEAD", "GET", "POST"])
 @login_required
 def announce(request):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     if request.method == "POST":
         form = AnnounceForm(request.POST)
         if form.is_valid():
@@ -921,6 +981,9 @@ def announce(request):
 @require_http_methods(["POST"])
 @login_required
 def users_deactivate(request):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     if not request.user.profile.is_admin:
         return redirect("main:people")
     if request.method == "POST":
@@ -939,6 +1002,9 @@ def users_deactivate(request):
 @require_http_methods(["POST"])
 @login_required
 def users_activate(request):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     if not request.user.profile.is_admin:
         return redirect("main:people")
     if request.method == "POST":
@@ -957,6 +1023,9 @@ def users_activate(request):
 @require_http_methods(["POST"])
 @login_required
 def users_adminify(request):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     if not request.user.profile.is_admin:
         return redirect("main:people")
     if request.method == "POST":
@@ -975,6 +1044,9 @@ def users_adminify(request):
 @require_http_methods(["POST"])
 @login_required
 def users_deadminify(request):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     if not request.user.profile.is_admin:
         return redirect("main:people")
     if request.method == "POST":
@@ -1020,6 +1092,9 @@ def billing_settings(request):
 @require_http_methods(["POST"])
 @login_required
 def account_delete(request):
+    if request.user.profile.is_admin and not request.user.profile.stripe_id:
+        return redirect("main:billing_setup")
+
     if not request.user.profile.is_admin:
         return redirect("main:settings_user")
     if request.method == "POST":
